% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectr.R
\name{chisqPgram}
\alias{chisqPgram}
\title{Calculate chi-square periodogram}
\usage{
chisqPgram(
  x,
  deltat,
  periodRange = c(18, 32),
  fair = TRUE,
  na.action = imputeTS::na_ma,
  dopar = TRUE
)
}
\arguments{
\item{x}{Numeric vector of measurements.}

\item{deltat}{Numeric value of the time interval between measurements.}

\item{periodRange}{Numeric vector of the minimum and maximum values of the
period, in units of \code{deltat}, for which to calculate the chi-squared
statistic.}

\item{fair}{Logical indicating whether to calculate the chi-squared statistic
based on the same number of blocks for each possible period. The
traditional calculation sets this variable to \code{FALSE}, which can lead to
discontinuities in the curve of chi-squared or p-value as a function of
period.}

\item{na.action}{Function specifying how to handle \code{NA} values in \code{x}.
Default is \verb{\link[imputeTS]\{na_ma\}()}, which imputes missing values by
weighted moving average.}

\item{dopar}{Logical indicating whether to run calculations in parallel, if
a parallel backend has already been set up, e.g., using
\verb{\link[doParallel]\{registerDoParallel\}()}.}
}
\value{
\code{data.table} with columns \code{period}, \code{chisq}, \code{df}, and \code{pval}.
}
\description{
Calculate the chi-square periodogram of a time series.
}
\examples{
library('data.table')

set.seed(1764)
deltat = 0.1
tau = 24
tt = seq(0, tau * 5, deltat)
x = sin(tt / tau * 2 * pi) + rnorm(length(tt))

specChisq = chisqPgram(x, deltat, dopar = FALSE)
peaksChisq = specChisq[which.min(pval)]

}
\seealso{
\verb{\link{spectrPgram}}
}
