% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_signature.R
\name{get.spectral.signature}
\alias{get.spectral.signature}
\title{Get spectral signatures from raw reference controls}
\usage{
get.spectral.signature(
  raw.fluor.control.paths,
  return.type = c("medians", "raw")
)
}
\arguments{
\item{raw.fluor.control.paths}{Character string; path(s) usually returned from \code{list.files(...,full.names=T,pattern=".fcs")}. For Cytek Aurora experiments, the raw reference controls are stored in: \code{"./.../Raw/Reference Group"}.}

\item{return.type}{Character string; one of:
\itemize{
\item \code{'medians'} -- return median expression values per-detector/per-control.
\item \code{'raw'} -- return raw expression values per-detector/per-control.
}}
}
\value{
a \code{spectr} object.
}
\description{
Using Cytek Aurora raw reference controls -- .fcs files containing emission values for all available detectors -- either median values per detector or the raw values themselves can be returned.
}
\examples{

extdata <- system.file("extdata",package="spectr")
raw.fluor.control.paths <- list.files(extdata,full.names=TRUE,pattern="Beads.*.fcs")

spectral.signatures<-get.spectral.signature(raw.fluor.control.paths)
spectral.signatures$data[]


}
