% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fcs_to._dt.R
\name{read.fcs.to.dt}
\alias{read.fcs.to.dt}
\title{Read .fcs files and concatenate (row-bind) into a \link[data.table]{data.table}.}
\usage{
read.fcs.to.dt(
  fcs.file.paths,
  colnames.syntatically.valid = TRUE,
  natural.order = TRUE,
  colnames.type = c("S", "S_N", "N"),
  cofactor = 5000,
  sample.id = "TUBENAME",
  keywords.to.factor = NULL
)
}
\arguments{
\item{fcs.file.paths}{Character string; path(s) usually returned from \code{list.files(...,full.names=T,pattern=".fcs")}.}

\item{colnames.syntatically.valid}{Logical; default \code{TRUE}. Syntactically valid names (letters, numbers and the underline character) are enforced by removing all dashes and spaces from column/parameter names.}

\item{natural.order}{Logical; default \code{TRUE}. Naturally order column names.}

\item{colnames.type}{Character string; one of:
\itemize{
\item \code{"S_N"} -- parameter columns are named by combining $PS and $PN, separated by an underscore. For Cytek Aurora instruments using SpectroFlo software, the default parameter naming scheme is as follows: $PS = stain (marker); $PN = detector/fluorophore.
\item \code{"S"} -- parameter columns are named by using only their respective $PS keyword value.
\item \code{"N"} -- parameter columns are named by using only their respective $PN keyword value.
}}

\item{cofactor}{Numeric; default \code{5000}. Any/all parameters with a \verb{$PnTYPE} of 'Unmixed_Fluorescence' will be transformed using \link{asinh} and the defined cofactor value (\code{asinh(x/cofactor)}).}

\item{sample.id}{Character string; the keyword label defined through \code{sample.id} (default \code{TUBENAME}) will be used to add respective keyword values as an identifier to the \code{data.table}.}

\item{keywords.to.factor}{Character string; if defined, keywords will be coerced to factor and appended to the returned data.table as factored columns.}
}
\value{
a list
}
\description{
Read .fcs files and concatenate (row-bind) into a \link[data.table]{data.table}.
}
\examples{

extdata <- system.file("extdata",package="spectr")
fcs.file.paths <- list.files(extdata,full.names=TRUE,pattern="BLOCK.*.fcs")

#trying to read mixed panels ("AIM","CYTOKINE") will cause an error due to $PS name conflicts
\dontrun{
read.fcs.to.dt(fcs.file.paths)
}

#"CYTOKINE" panel
fcs.file.paths<-grep("CYTOKINE",fcs.file.paths,value=TRUE)

dt<-read.fcs.to.dt(fcs.file.paths)

names(dt)
sapply(c("data","parameters","parameters.non"),function(i){head(dt[[i]])})
head(dt$spill[[1]])

#add additional (factored) keywords
dt<-read.fcs.to.dt(fcs.file.paths,keywords.to.factor=c('$PROJ','$DATE'))

dt$data

}
