% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectr-package.R
\docType{package}
\name{spectr-package}
\alias{spectr}
\alias{spectr-package}
\title{spectr: Full Specturm Flow Cytometry Analysis Tools}
\description{
Analysis and visualization of full spectrum flow cytometry .fcs data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nlaniewski/spectr}
  \item Report bugs at \url{https://github.com/nlaniewski/spectr/issues}
}

}
\author{
\strong{Maintainer}: Nathan Laniewski \email{nathan_laniewski@urmc.rochester.edu}

}
\keyword{internal}
