% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectr.R
\name{spectr}
\alias{spectr}
\title{Calculate FFT-based periodogram}
\usage{
spectr(
  x,
  deltat,
  periodRange = c(18, 32),
  pad = 50,
  na.action = imputeTS::na_ma,
  ...
)
}
\arguments{
\item{x}{Numeric vector of measurements.}

\item{deltat}{Numeric value of the time interval between measurements.}

\item{periodRange}{Numeric vector of the minimum and maximum values of the
period to consider, in units of \code{deltat}.}

\item{pad}{Numeric value of the proportion of the length of \code{x} by which to
pad \code{x} with zeros.}

\item{na.action}{Function specifying how to handle \code{NA} values in \code{x}.
Default is \code{\link[imputeTS:na_ma]{imputeTS::na_ma()}}, which imputes missing values by weighted
moving average.}

\item{...}{Other arguments passed to \code{\link[stats:spec.pgram]{stats::spec.pgram()}}.}
}
\value{
\code{data.table} of the estimated spectral density, with columns \code{period}
(in the same units as \code{deltat}) and \code{power}.
}
\description{
Calculate the periodogram of a time series using a fast Fourier transform.
}
\examples{
library('data.table')

set.seed(1764)
deltat = 0.1
tau = 24
tt = seq(0, tau * 5, deltat)
x = sin(tt / tau * 2 * pi) + rnorm(length(tt))

spec = spectr(x, deltat)
specPeak = spec[which.max(power)]

}
