% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_signature.R
\name{spectral.signature.plot}
\alias{spectral.signature.plot}
\title{Plot a spectral signature}
\usage{
spectral.signature.plot(raw.fluor.control.paths, subtract.universal = TRUE)
}
\arguments{
\item{raw.fluor.control.paths}{Character string; path(s) usually returned from \code{list.files(...,full.names=T,pattern=".fcs")}. For Cytek Aurora experiments, the raw reference controls are stored in: \code{"./.../Raw/Reference Group"}.}

\item{subtract.universal}{Logical; default \code{TRUE}. Per-detector background fluorescence will be subtracted using a universal negative control, with respect to control-type (beads or cells).}
}
\value{
A list containing \link[ggplot2]{ggplot2} objects.
}
\description{
Using Cytek Aurora raw reference controls -- .fcs files containing emission values for all available detectors -- median values per detector/per-control are calculated and then used to plot a 'trace' line representing a spectral signature.
}
\examples{

extdata <- system.file("extdata",package="spectr")
raw.fluor.control.paths <- list.files(extdata,full.names=TRUE,pattern="Beads.*.fcs")

p<-spectral.signature.plot(raw.fluor.control.paths)
p[]

}
