% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectr.R
\name{spectrAlpha}
\alias{spectrAlpha}
\title{Calculate approximate alpha from a circadian time-course}
\usage{
spectrAlpha(time, activity, tau, thresh, frac = 0.9)
}
\arguments{
\item{time}{Numeric vector of time.}

\item{activity}{Numeric vector of activity measurements.}

\item{tau}{Numeric value of the period of the time series.}

\item{thresh}{Numeric value of the activity threshold. If \code{NULL}, the
\code{activity} vector is not binarized.}

\item{frac}{Numeric value of the fraction of activity above threshold (or
fraction of overall activity, if \code{thresh} is \code{NULL}) to capture.}
}
\value{
\code{data.table} with columns \code{onset}, \code{offset}, and \code{width}, all in
units of fraction of \code{tau}.
}
\description{
TODO.
}
\examples{
library('data.table')

# TODO

}
\seealso{
\verb{\link{spectrPgram}}, \verb{\link{spectrPeaks}}
}
