% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectr.R
\name{spectrPeaks}
\alias{spectrPeaks}
\title{Find peaks in periodogram}
\usage{
spectrPeaks(spec, nPeaks = 1L, splineDf = 3L, ...)
}
\arguments{
\item{spec}{\code{data.table} with at least columns \code{freq} and \code{power}.}

\item{nPeaks}{Integer of the maximum number of peaks to return, sorted by
decreasing \code{power}.}

\item{splineDf}{Numeric value of degrees of freedom for the natural cubic
spline used to fit the periodogram around each peak, passed to
\code{\link[splines]{ns}()}.}

\item{...}{Other arguments passed to \code{\link[pracma]{findpeaks}()}.}
}
\value{
\code{data.table} of periodogram peaks, with columns \code{freq}, \code{period}, and
\code{power}.
}
\description{
Find local maxima and use spline interpolation to estimate one or more peaks
in the spectral density.
}
\examples{
library('data.table')

set.seed(1764)
deltat = 0.1
tau = 24
tt = seq(0, tau * 5, deltat)
x = sin(tt / tau * 2 * pi) + rnorm(length(tt))

spec = spectrPgram(x, deltat)
peaks = spectrPeaks(spec[period \%between\% (tau + c(-4, 4))])

}
\seealso{
\code{\link{spectrPgram}}
}
