% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectr.R
\name{spectrPgram}
\alias{spectrPgram}
\title{Calculate periodogram}
\usage{
spectrPgram(x, deltat, pad = 20, na.action = imputeTS::na_ma, ...)
}
\arguments{
\item{x}{Numeric vector of measurements.}

\item{deltat}{Numeric value of the time interval between measurements.}

\item{pad}{Numeric value of the proportion of the length of \code{x} by which to
pad \code{x} with zeros, passed to \verb{\link[stats]\{spec.pgram\}()}.}

\item{na.action}{Function specifying how to handle \code{NA} values in \code{x},
passed to \verb{\link[stats]\{spec.pgram\}()}. Default is
\verb{\link[imputeTS]\{na_ma\}()}, which imputes missing values by weighted
moving average.}

\item{...}{Other arguments passed to \verb{\link[stats]\{spec.pgram\}()}.}
}
\value{
\code{data.table} of the estimated spectral density, with columns \code{period}
(in the same units as \code{deltat}) and \code{power}.
}
\description{
Calculate the periodogram of a time series using a fast Fourier transform.
}
\examples{
library('data.table')

set.seed(1764)
deltat = 0.1
tau = 24
tt = seq(0, tau * 5, deltat)
x = sin(tt / tau * 2 * pi) + rnorm(length(tt))

spec = spectrPgram(x, deltat)
peaks = spectrPeaks(spec[period \%between\% (tau + c(-4, 4))])

}
\seealso{
\verb{\link{spectrPeaks}}
}
