% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectr.R
\name{chisqPgram}
\alias{chisqPgram}
\title{Calculate chi-square periodogram}
\usage{
chisqPgram(x, deltat, na.action = imputeTS::na_ma, dopar = TRUE)
}
\arguments{
\item{x}{Numeric vector of measurements.}

\item{deltat}{Numeric value of the time interval between measurements.}

\item{na.action}{Function specifying how to handle \code{NA} values in \code{x}.
Default is \verb{\link[imputeTS]\{na_ma\}()}, which imputes missing values by
weighted moving average.}

\item{dopar}{Logical indicating whether to run calculations in parallel, if
a parallel backend has already been set up, e.g., using
\verb{\link[doParallel]\{registerDoParallel\}()}.}
}
\value{
\code{data.table} with columns \code{period} and \code{chisq}.
}
\description{
Calculate the chi-square periodogram of a time series.
}
\examples{
library('data.table')

set.seed(1764)
deltat = 0.1
tau = 24
tt = seq(0, tau * 5, deltat)
x = sin(tt / tau * 2 * pi) + rnorm(length(tt))

specChisq = chisqPgram(x, deltat, dopar = FALSE)
peaksChisq = specChisq[period \%between\% (tau + c(-4, 4))][which.max(chisq)]

}
\seealso{
\verb{\link{spectrPgram}}
}
