% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.signature.R
\name{spectral.signature}
\alias{spectral.signature}
\title{Plot a normalized spectral signature}
\usage{
spectral.signature(
  raw.fluor.control.path,
  raw.unstained.control.path,
  transform.type = "asinh",
  cofactor = 5000,
  top.n = 200
)
}
\arguments{
\item{raw.fluor.control.path}{Character string; file path to a spectral flow cytometry reference (fluor) control.}

\item{raw.unstained.control.path}{Character string; file path to a spectral flow cytometry reference (unstained) control.}

\item{transform.type}{Character string; default and only current choice is 'asinh' for arc-sine transformation of fluorescence detectors.}

\item{cofactor}{Numeric; default 5000. Defines the cofactor value used during arc-sine transformation of fluorescence detectors.}

\item{top.n}{Numeric; default 200. Defines the top number of expressing events (peak detector) to use when generating the normalized median values.}
}
\value{
A \link[ggplot2]{ggplot} object.
}
\description{
Plot a normalized spectral signature for spectral flow cytometry reference controls.
}
\examples{
raw.fluor.control.path<-list.files(
system.file("extdata",package="spectr"),
full.names = TRUE,
pattern = "RB744"
)

raw.unstained.control.path<-list.files(
system.file("extdata",package="spectr"),
full.names = TRUE,
pattern = "Unstained"
)

spectral.signature(
raw.fluor.control.path,
raw.unstained.control.path
)

raw.fluor.control.path<-list.files(
system.file("extdata",package="spectr"),
full.names = TRUE,
pattern = "PerCPFire806"
)

spectral.signature(
raw.fluor.control.path,
raw.unstained.control.path
)
}
